agora-analyst-python/requirements.txt
# Production Dependencies for Kritis 4.0
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0
numpy>=1.24.0,<2.0.0
pgvector>=0.2.0,<1.0.0
requests>=2.31.0,<3.0.0